+++ desktop/php/jMQTT.php
@@ -27,6 +27,9 @@ span.hiddenAsCard i.fas.fa-sign-in-alt				{ margin-right:10px;vertical-align:top
 span.hiddenAsCard i.fas.status-circle				{ margin-right:6px; }
 textarea.form-control.input-sm.modifiedVal			{ color: darkorange!important; font-weight: bold!important; }
 div.eqLogicDisplayCard[jmqtt_type="broker"]			{ background: rgba(248, 216, 0, 0.25)!important; }
+textarea.eqLogicAttr.form-control.blured			{ filter: blur(4px); }
+textarea.eqLogicAttr.form-control.blured:hover		{ filter: none; }
+textarea.eqLogicAttr.form-control.cert				{ font-family: "CamingoCode", monospace; height: 90px; }
 </style>
 
 <?php